requirements.txt
# PyTorch and deep learning
torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0

# Data handling and processing
numpy>=1.21.0
pandas>=1.3.0
pillow>=8.3.0

# Machine learning utilities
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Progress bars
tqdm>=4.62.0

# Dataset download
kaggle>=1.5.12

# Configuration
pyyaml>=6.0

# Jupyter notebooks
jupyter>=1.0.0
notebook>=6.4.0
ipykernel>=6.0.0

# Image processing
opencv-python>=4.5.0

# Additional utilities
requests>=2.25.0